added FAB to test app as well as activity for debug. improved test app.
